fix TimeMultiplier from 0-24

+++ backend/src/main/java/grimsi/accservermanager/backend/entity/Session.java
@@ -4,14 +4,13 @@ import grimsi.accservermanager.backend.enums.SessionType;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 
 @Getter
 @Setter
 public class Session {
     private int hourOfDay;
     private int dayOfWeekend;
-    private BigDecimal timeMultiplier;
+    private int timeMultiplier;
     private SessionType sessionType;
     private int sessionDurationMinutes;
 }
